Fixed collection MaxTTL to be uint32 to match manifest.

+++ cbmgmtx/mgmt.go
@@ -346,7 +346,7 @@ type CreateCollectionOptions struct {
 	BucketName     string
 	ScopeName      string
 	CollectionName string
-	MaxTTL         uint64
+	MaxTTL         uint32
 }
 
 func (h Management) CreateCollection(
